# SavePoint/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SavePoint API",
    "version": "1.0.0",
    "description": "API REST do SavePoint: catálogo de jogos, comparação de preços entre lojas, histórico de preços, avaliações de usuários e lista de desejos."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/jogos/{id}/requisitos-minimos": {
      "get": {
        "tags": [
          "Requisitos mínimos"
        ],
        "summary": "Buscar requisitos mínimos do jogo",
        "operationId": "requisitosMinimos_buscarPorJogoId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequisitosMinimos"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Requisitos mínimos"
        ],
        "summary": "Cadastrar requisitos mínimos do jogo",
        "operationId": "requisitosMinimos_cadastrar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequisitosMinimos"
              },
              "example": {
                "processador": "Intel Core i5-2500K 3.3 GHz",
                "memoria": "8 GB RAM",
                "placaDeVideo": "NVIDIA GeForce GTX 660",
                "sistemaOperacional": "Windows 10 64-bit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequisitosMinimos"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou jogo já possui requisitos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Jogo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Requisitos mínimos"
        ],
        "summary": "Atualizar requisitos mínimos do jogo",
        "operationId": "requisitosMinimos_atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequisitosMinimos"
              },
              "example": {
                "processador": "Intel Core i5-2500K 3.3 GHz",
                "memoria": "8 GB RAM",
                "placaDeVideo": "NVIDIA GeForce GTX 660",
                "sistemaOperacional": "Windows 10 64-bit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requisitos atualizados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequisitosMinimos"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Requisitos mínimos"
        ],
        "summary": "Excluir requisitos mínimos do jogo",
        "operationId": "requisitosMinimos_deletar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
    "/jogos/{id}/resumo": {
      "post": {
        "tags": [
          "Resumo de avaliações"
        ],
        "summary": "Gerar resumo de avaliações do jogo",
        "operationId": "resumoAvaliacao_gerarResumo",
        "description": "Calcula nota média, total e percentual de recomendação. Se já existir um resumo, ele é recalculado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Resumo gerado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoAvaliacao"
                }
              }
            }
          },
          "404": {
            "description": "Jogo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Resumo de avaliações"
        ],
        "summary": "Buscar resumo de avaliações do jogo",
        "operationId": "resumoAvaliacao_buscarResumo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoAvaliacao"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Resumo de avaliações"
        ],
        "summary": "Atualizar resumo de avaliações do jogo",
        "operationId": "resumoAvaliacao_atualizarResumo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumoAvaliacao"
              },
              "example": {
                "resumoGeradoIA": "Jogo muito elogiado pela narrativa e pelo mundo aberto.",
                "principaisElogios": [
                  "Narrativa envolvente",
                  "Trilha sonora"
                ],
                "principaisCriticas": [
                  "Bugs no lancamento"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resumo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoAvaliacao"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/jogos/{jogoId}/resumo/{resumoId}": {
      "delete": {
        "tags": [
          "Resumo de avaliações"
        ],
        "summary": "Excluir resumo de avaliações",
        "operationId": "resumoAvaliacao_deletarResumo",
        "parameters": [
          {
            "name": "jogoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "resumoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RequisitosMinimos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processador": {
            "type": "string"
          },
          "memoria": {
            "type": "string"
          },
          "placaDeVideo": {
            "type": "string"
          },
          "sistemaOperacional": {
            "type": "string"
          }
        }
      },
      "ResumoAvaliacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "notaMedia": {
            "type": "number",
            "format": "double"
          },
          "totalAvaliacoes": {
            "type": "integer",
            "format": "int32"
          },
          "percentualRecomendacao": {
            "type": "number",
            "format": "double"
          },
          "principaisElogios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "principaisCriticas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "resumoGeradoIA": {
            "type": "string"
          },
          "dataGeracao": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ErroResposta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "erro": {
            "type": "string"
          },
          "mensagem": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}